chore(server): rename object metadata service v2
