+++ packages/twenty-server/src/engine/core-modules/open-api/open-api.service.spec.ts
@@ -5,7 +5,7 @@ import { FeatureFlagService } from 'src/engine/core-modules/feature-flag/service
 import { OpenApiService } from 'src/engine/core-modules/open-api/open-api.service';
 import { TwentyConfigService } from 'src/engine/core-modules/twenty-config/twenty-config.service';
 import { WorkspaceManyOrAllFlatEntityMapsCacheService } from 'src/engine/metadata-modules/flat-entity/services/workspace-many-or-all-flat-entity-maps-cache.service';
-import { ObjectMetadataServiceV2 } from 'src/engine/metadata-modules/object-metadata/object-metadata-v2.service';
+import { ObjectMetadataServiceV2 } from 'src/engine/metadata-modules/object-metadata/object-metadata.service';
 
 describe('OpenApiService', () => {
   let service: OpenApiService;
